Sir_fudge/requirements.txt
flask
flask-cors
openai-whisper
torch
setuptools-rust
werkzeug
gunicorn